修改 customer 客户信息

+++ ses-service/ses-service-hub/src/main/java/com/redescooter/ses/service/hub/source/operation/dm/OpeCustomer.java
@@ -64,7 +64,7 @@ public class OpeCustomer implements Serializable {
      */
     @TableField(value = "city")
     @ApiModelProperty(value = "城市")
-    private Integer city;
+    private Long city;
 
     /**
      * 区域
